+++ src/ElectronDownloadManagerMock.ts
@@ -1,0 +1,23 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
+import { DownloadParams, IElectronDownloadManager } from './types'
+
+/**
+ * Mock version of ElectronDownloadManager
+ * that can be used for testing purposes
+ */
+export class ElectronDownloadManagerMock implements IElectronDownloadManager {
+  download(_params: DownloadParams): string {
+    return 'mock-download-id'
+  }
+
+  cancelDownload(_id: string): void {}
+
+  pauseDownload(_id: string): void {}
+
+  resumeDownload(_id: string): void {}
+
+  getActiveDownloadCount(): number {
+    return 0
+  }
+}