+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0 (2024-03-21)
+
+- Add `ElectronDownloadManagerMock` for use in tests
+- Add `getActiveDownloadCount()` to get the number of active downloads
+- Add `showBadge` option to `download()` to show a badge on the dock icon
+
 # 1.0.2 (2024-03-21)
 
 `download()` now returns the `id` of the download